+++ core/include/dlvaracc.h
@@ -77,7 +77,7 @@
 #define LOCEXISTS(m,a)    (m##_locals_fixup[a].size != -1)
 #define GLOEXISTS(m,a)    (m##_globals_fixup[a].data_offset)
 
-#define LOCADDR(m,a,b)    ((uint8_t *)((a)->locdata) + (uint64_t)m##_locals_fixup[b].data_offset)
+#define LOCADDR(m,a,b)    ((uint8_t *)(intptr_t)((a)->locdata) + (uint64_t)(intptr_t)m##_locals_fixup[b].data_offset)
 #define GLOADDR(m,a)      (m##_globals_fixup[a].data_offset)
 
 #define LOCQWORD(m,a,b)   (*(uint64_t *)LOCADDR(m,a,b))
